requirements.txt
gspread==5.7.2
numpy==1.23.2
pandas==1.4.3
Pillow==9.4.0
streamlit==1.16.0
streamlit_authenticator==0.2.1
openpyxl==3.0.10
